+++ jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/kafka/producer/DisabledTopicsStringProducerInterceptor.java
@@ -1,11 +1,6 @@
 package us.dot.its.jpo.ode.kafka.producer;
 
-import java.util.Map;
-import java.util.Set;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.kafka.clients.producer.ProducerInterceptor;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.clients.producer.RecordMetadata;
 import us.dot.its.jpo.ode.kafka.OdeKafkaProperties;
 
 /**
@@ -19,35 +14,9 @@ import us.dot.its.jpo.ode.kafka.OdeKafkaProperties;
  */
 @Slf4j
 public class DisabledTopicsStringProducerInterceptor
-    implements ProducerInterceptor<String, String> {
-
-  private final Set<String> disabledTopics;
+    extends AbstractDisabledTopicsProducerInterceptor<String, String> {
 
   public DisabledTopicsStringProducerInterceptor(OdeKafkaProperties odeKafkaProperties) {
-    this.disabledTopics = odeKafkaProperties.getDisabledTopics();
-  }
-
-  @Override
-  public ProducerRecord<String, String> onSend(ProducerRecord<String, String> producerRecord) {
-    if (disabledTopics.contains(producerRecord.topic())) {
-      throw new DisabledTopicException(producerRecord.topic());
-    }
-    return producerRecord;
-  }
-
-  @Override
-  public void onAcknowledgement(RecordMetadata recordMetadata, Exception e) {
-    log.debug("Acknowledged message with offset {} on partition {}", recordMetadata.offset(),
-        recordMetadata.partition());
-  }
-
-  @Override
-  public void close() {
-    log.debug("Closing StringProducerInterceptor");
-  }
-
-  @Override
-  public void configure(Map<String, ?> map) {
-
+    super(odeKafkaProperties.getDisabledTopics());
   }
 }
